requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
black==23.3.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
Chroma==0.2.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.20
commonmark==0.9.1
confection==0.0.4
cryptography==40.0.2
cymem==2.0.7
dataclasses-json==0.5.7
DAWG-Python==0.7.2
Deprecated==1.2.13
deptry==0.11.0
dnspython==2.3.0
docopt==0.6.2
duckdb==0.7.1
ecdsa==0.18.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
fastapi==0.95.1
filelock==3.12.0
frozenlist==1.3.3
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.171
langcodes==3.3.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
minio==7.1.14
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pathy==0.10.1
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
posthog==3.0.0
preshed==3.0.8
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.15.1
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
PyMuPDF==1.22.1
pypandoc==1.11
pypdf==3.8.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
ru-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.5.0/ru_core_news_lg-3.5.0-py3-none-any.whl#sha256=5e76e86d4c5008296cf77c9080269908ea4097794585f6e956b3462ad5e8c037
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.5.0/ru_core_news_sm-3.5.0-py3-none-any.whl#sha256=bedf427ab4f522137822d52c50ebdf9edeef31fdb808ed4776ce1596aa789704
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
srsly==2.4.6
sse-starlette==1.6.0
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.6.8
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
Wand==0.6.11
wasabi==1.1.1
watchfiles==0.19.0
websockets==11.0.2
wrapt==1.14.1
XlsxWriter==3.1.1
yarl==1.8.2
zstandard==0.21.0
